% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_name_list.R
\docType{data}
\name{field_name}
\alias{field_name}
\title{field_name}
\format{A string array that contains the column names in hedonics table}
\usage{
data(field_name)
}
\description{
field_name
}
\keyword{hedonics}
