% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zillow.R
\name{get_from_db_fips}
\alias{get_from_db_fips}
\title{get_from_db_fips}
\usage{
get_from_db_fips(fips, columns = "*", database_name = "zillow_2017_nov",
  host_ip = "141.142.209.139", append = TRUE)
}
\arguments{
\item{fips}{A vector of fips}

\item{columns}{A vector of column names to export. Default to all columns (i.e. "*").}

\item{database_name}{A string indicating the database name}

\item{host_ip}{A string indicating the ip address of the database VM}

\item{append}{If append is true, return a single data.frame with rows appended, otherwise a
list of data.frames from each state.}
}
\value{
A data.frame including all data from the given fips.
}
\description{
This function gets a data.frame including all data from the given fips. Use previous
             function.
}
\examples{
data <- get_from_db_fips("10001")
data <- get_from_db_fips(c("01001","17019"))
}
