% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zillow.R
\name{get_from_db_state}
\alias{get_from_db_state}
\title{get_from_db_state}
\usage{
get_from_db_state(states_abbr, columns = "*", max_num_recs = -1,
  database_name = "zillow_2017_nov", host_ip = "141.142.209.139",
  append = TRUE)
}
\arguments{
\item{states_abbr}{A vector of states abbreviation}

\item{columns}{A vector of column names to export. Default to all columns (i.e. "*").}

\item{max_num_recs}{An integer indicating the maximum number of records to select,
-1 indicating all (by default)}

\item{database_name}{A string indicating the database name}

\item{host_ip}{A string indicating the ip address of the database VM}

\item{append}{If append is true, return a single data.frame with rows appended, otherwise a
list of data.frames from each state.}
}
\value{
If input is a single state, return a single data.frame. If input is a vector and append = F,
        return a list of data.frames, in the same order of that in states_abbr, otherwise the result
        is appended into one whole data.frame.
}
\description{
This function gets a list of data.frames from the database, according to the given
             states (through abbreviation).
}
\examples{
 data <- get_from_db_state("sd")
 data <- get_from_db_state(c("sD","Ne"), append=FALSE)
 data <- get_from_db_state("ca",columns=c("rowid","transid"))
 data <- get_from_db_state("sd",max_num_recs=100)
}
