% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zillow.R
\name{get_from_db_usr}
\alias{get_from_db_usr}
\title{get_from_db_usr}
\usage{
get_from_db_usr(query, database_name = "zillow_2019_sep",
  host_ip = "141.142.209.139")
}
\arguments{
\item{query}{A string specifying the query sent to database}

\item{database_name}{A string indicating the database name}

\item{host_ip}{A string indicating the ip address of the database VM}
}
\value{
A data.frame returned by the given query.
}
\description{
This function gets from database according to a user-specified query.
}
\examples{
# Select single field from a given table
data <- get_from_db_usr("SELECT loadid FROM hedonics_new.sd_hedonics_new")

# Select multiple fields from a given table
data <- get_from_db_usr("SELECT loadid, transid FROM hedonics_new.sd_hedonics_new")

# Select limited rows from a given table
data <- get_from_db_usr("SELECT * FROM hedonics_new.sd_hedonics_new limit 10")

# Select a 'bounding box' from a given table
data <- get_from_db_usr("SELECT * FROM hedonics_new.sd_hedonics_new WHERE (propertyaddresslatitude BETWEEN 44.35 AND 44.36) AND (propertyaddresslongitude BETWEEN -98.22 AND -98.21)")

# Select from a list for a column (columns) in a given table
get_from_db_usr("SELECT * FROM hedonics_new.sd_hedonics_new WHERE county IN ('BEADLE', 'UNION')")

# Get the number of records (rows) in a given table
get_from_db_usr("SELECT count(*) FROM hedonics_new.sd_hedonics_new")

# Get the range of a specific field
get_from_db_usr("SELECT min(propertyaddresslatitude), max(propertyaddresslatitude) FROM hedonics_new.sd_hedonics_new")
}
