% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_state_fips_table.R
\name{get_state_county_by_names}
\alias{get_state_county_by_names}
\title{get_state_county_by_names}
\usage{
get_state_county_by_names(sc)
}
\arguments{
\item{sc}{A data.frame containing state abbreviation and county names}
}
\value{
A data.frame with first column as state abbreviation, second column as county name,
        both in lower case, and third column fips.
}
\description{
This function convert a data.frame of state abbr. and county names to corresponding state-county pairs + fips
}
